hardcoded an initial idea. Added the required Google Calendar dlls Process reading code is commented out, and calendar code is hardcoded. The concept should work just fine.  Calendar entries will save.  It needs to be fleshed out.
The initial concept is to create a console program that sits in the background and passively logs the specified process.
The processes to monitor will probably be laid out as entries in an XML file with some sort of action associated with them. At present, it will simply be: log to calendar.  Later there will be a feature to start up processes based upon entries placed in the calendar.

As far as an error log, I'd like to make the errors show up as calendar entries.  If connection can't be made to the calendar, then they will be logged in the system event log.