Added XML Functionality.
Refactored code again.  

This is *almost* the program I set out to write, but now we need to clean up so we can get v1.0 done.

later it can be made more robust (like reporting on specific tabs in chrome, recovering from errors, making the logging more robust, etc.

Whats left is to turn it into a service, make a *login* screen (which I can steal from my other project, ChessMangler), and let the user choose their calendar from it. (but login info will not be saved for v1.0, but the calendar might..)