Bug fixed.
An entry will now be saved to the calendar for each instance of the program that is closed.  This was actually causing a problem when you closed a single tab in chrome.  It would crash the Program.

The bad news is that multiple tabs in chrome show up as separate instances. As it may be desired to see when some tabs are closed, thisneeds to be made into an option.